+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2021-10-13
+
+* Added Naga compiler v0.7.0 (#101)
+
 ## 2021-09-25
 
 * Added Metal compiler 2.0 beta 5